refactored Product Tags to be in its own package
