+++ src/main/java/com/revature/shoply/productTags/controllers/ProductTagController.java
@@ -1,10 +1,8 @@
-package com.revature.shoply.controllers;
+package com.revature.shoply.productTags.controllers;
 
 import com.revature.shoply.models.Product;
-import com.revature.shoply.models.Tag;
 import com.revature.shoply.product.service.ProductService;
 import com.revature.shoply.product.service.ProductTagService;
-import com.revature.shoply.services.TagService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
